requirements.txt
langchain==0.0.346
pgvector==0.2.3
protobuf==4.24.4
psycopg2-binary==2.9.9
sentence_transformers==2.2.2
unstructured[all,pdf,pptx]>=0.10.22

python-dotenv==1.0.0
pandas==2.1.1

tiktoken==0.5.1
uvicorn[standard]
pydantic
fastapi
dependency-injector
fastapi_health
asyncpg
orjson
langfuse==1.8.0

uvicorn[standard]==0.24.0.post1
pydantic==2.5.2
fastapi==0.105.0
dependency-injector==4.41.0
fastapi_health==0.4.0
asyncpg==0.29.0
orjson==3.9.10
loguru==0.7.2
ansi2html==1.9.1
arize-phoenix==1.9.0